requirements.txt
py-cord==2.0.0b7
python-dotenv==0.20.0
